requirements.txt
requests
beautifulsoup4
Pillow
numpy
opencv-python